services/observability_service/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.27,<1.0
redis>=5.2,<6.0
psycopg[binary]>=3.2,<4.0
prometheus-client>=0.21,<1.0
python-dotenv>=1.0,<2.0
